Rename render.rs to bitmap.rs for naming consistency with outline.rs
Both module names are now nouns describing their domain.
