+++ src/bitmap.rs
